[Propeller] Match the hash value of BB to adapt to source drift
Add the BB Hash to the generated `.llvm_bb_addr_map` section. The
Propeller tool reads it and write to the output file. When building
updated code, llvm matches the file with the current BB Hash and
generates BBCluster based on the Hash matching results. This feature is
controlled by the `bbsections-match-hash` option and is disabled by
default.